+++ src/lib/types/PostSchema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { postSchema } from '@/schemas';
+import { postSchema } from '@/lib/schemas';
 
 type PostSchema = z.infer<typeof postSchema>;
